email-notify/requirements.txt
fastapi
uvicorn
jinja2
python-dotenv
pytest
